Renamed fixtures file
