+++ tests/functional/adapter/bigquery_test/incremental_strategy_fixtures.py
